Fixed most warnings. (we're at /W3 now because microsoft won't suppress valid c++11 behavior C4201) Readded Subscriber support in Actor3D. Fixed typo in Dispatcher comment

+++ source/util/timer.cpp
@@ -14,9 +14,11 @@ void Timer::Reset() {
 }
 
 void Timer::Update() {
-    delta_time_ = std::chrono::duration_cast<std::chrono::duration<double>>(
-                    std::chrono::high_resolution_clock::now() - start_ticks_
-                  ).count();
+    delta_time_ = static_cast<float32>(
+                    std::chrono::duration_cast<std::chrono::duration<double>>(
+                      std::chrono::high_resolution_clock::now() - start_ticks_
+                    ).count()
+                  );
     start_ticks_ = std::chrono::high_resolution_clock::now();
 }
